third_party/WebKit/LayoutTests/http/tests/flags/mixed-mode.html

<!DOCTYPE html>
<title>Origin Flags API: Mixed Modes</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

promise_test(async t => {
  let unblock;
  const blocked = new Promise(r => { unblock = r; });

  const granted = [];

  // These should be granted immediately, and held until unblocked.
  await Promise.all([
    requestFlag('a', 'shared')
      .then(f => { granted.push('as1'); f.waitUntil(blocked); }),
    requestFlag('a', 'shared')
      .then(f => { granted.push('as2'); f.waitUntil(blocked); }),
    requestFlag('a', 'shared')
      .then(f => { granted.push('as3'); f.waitUntil(blocked); })
  ]);

  // This should be blocked.
  const excl = requestFlag('a', 'exclusive')
    .then(f => { granted.push('ae'); return f; });

  // This should be granted immediately (different scope).
  await requestFlag('b', 'shared').then(f => { granted.push('bs'); });
  assert_array_equals(granted, ['as1', 'as2', 'as3', 'bs']);

  unblock();

  // Now the blocked request can be granted.
  const flag = await excl;
  assert_equals(flag.mode, 'exclusive');

  assert_array_equals(granted, ['as1', 'as2', 'as3', 'bs', 'ae']);

}, 'Flag requests are granted in order');

</script>
